theme color of app

+++ app/src/main/java/site/xiaozk/dailyfitness/page/body/BodyDetailPage.kt
@@ -4,18 +4,13 @@ import androidx.compose.foundation.ExperimentalFoundationApi
 import androidx.compose.foundation.clickable
 import androidx.compose.foundation.combinedClickable
 import androidx.compose.foundation.layout.Box
-import androidx.compose.foundation.layout.Row
-import androidx.compose.foundation.layout.fillMaxHeight
 import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.layout.fillMaxWidth
-import androidx.compose.foundation.layout.height
 import androidx.compose.foundation.layout.padding
-import androidx.compose.foundation.layout.width
 import androidx.compose.foundation.lazy.LazyColumn
 import androidx.compose.foundation.lazy.items
-import androidx.compose.material.AlertDialog
+import androidx.compose.material3.AlertDialog
 import androidx.compose.material3.Card
-import androidx.compose.material3.MaterialTheme
 import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.collectAsState
@@ -70,40 +65,27 @@ fun BodyDetailPage() {
         val dismiss by rememberUpdatedState(newValue = { deleteBodyDialog = null })
         AlertDialog(
             onDismissRequest = dismiss,
-            buttons = {
-                Row(
+            confirmButton = {
+
+                Text(
+                    text = "删除",
                     modifier = Modifier
-                        .fillMaxWidth()
-                        .height(40.dp)
-                ) {
-                    Text(
-                        text = "取消",
-                        modifier = Modifier
-                            .weight(1f)
-                            .fillMaxHeight()
-                            .clickable { dismiss() },
-                        textAlign = TextAlign.Center
-                    )
-                    Box(
-                        modifier = Modifier
-                            .fillMaxHeight()
-                            .width(1.dp)
-                    )
-                    Text(
-                        text = "删除",
-                        modifier = Modifier
-                            .weight(1f)
-                            .fillMaxHeight()
-                            .clickable {
-                                deleteBodyDialog?.let {
-                                    viewModel.deleteBodyDetail(it)
-                                }
-                                dismiss()
-                            },
-                        color = MaterialTheme.colorScheme.error,
-                        textAlign = TextAlign.Center
-                    )
-                }
+                        .clickable {
+                            deleteBodyDialog?.let {
+                                viewModel.deleteBodyDetail(it)
+                            }
+                            dismiss()
+                        },
+                    textAlign = TextAlign.Center
+                )
+            },
+            dismissButton = {
+                Text(
+                    text = "取消",
+                    modifier = Modifier
+                        .clickable { dismiss() },
+                    textAlign = TextAlign.Center
+                )
             },
             title = {
                 Text(text = "确认删除？")
